lint

+++ mapper/block.go
@@ -12,7 +12,7 @@ func BlockMetadata(block *corethTypes.Block) map[string]interface{} {
 		"gas_used":   hexutil.EncodeUint64(block.GasUsed()),
 		"difficulty": block.Difficulty(),
 		"nonce":      block.Nonce(),
-		"size":       hexutil.EncodeUint64(uint64(block.Size())),
+		"size":       hexutil.EncodeUint64(block.Size()),
 	}
 	if block.BaseFee() != nil {
 		meta["base_fee"] = hexutil.EncodeBig(block.BaseFee())
